Module Deriver.Located

Parameters

A : Ppxlib.Ast_builder.S

Signature

val derive_str : ?⁠name:string -> ?⁠lib:Ppxlib.expression -> (Ppxlib.rec_flag * Ppxlib.type_declaration list) -> Ppxlib.structure_item list

Deriver for Irmin generics.

?name: overrides the default name of the generated generic;

?lib: overrides the default location for the primitive Irmin generics. ~lib:None will assume that the generics are available in the same namespace.

val derive_sig : ?⁠name:string -> ?⁠lib:Ppxlib.expression -> (Ppxlib.rec_flag * Ppxlib.type_declaration list) -> Ppxlib.signature_item list

Deriver for Irmin generic type signatures.

Optional arguments have the same meaning as in derive_str.