Module Ppx_irmin_lib__Algebraic.Located

Parameters

S : Ppxlib.Ast_builder.S

M : Ppx_irmin_lib__.Monad.S

Signature

type nonrec record_field_repr = {

field_name : string;

field_generic : Ppxlib.expression;

}

and variant_case_repr = {

case_name : string;

case_cons : (Ppxlib.expression * int) option;

}

type (_, _) typ =

| Record : (Ppxlib.label_declaration, record_field_repr) typ

| Variant : (Ppxlib.constructor_declaration, variant_case_repr) typ

| Polyvariant : (Ppxlib.row_field, variant_case_repr) typ

The algebraic datatypes supported by this module, parameterised by:

'a: the subcomponent type of the algebraic type

'b: a generic representation of the subcomponent type necessary to derive the composite generic

module M = M

val encode : ('a, 'b) typ -> subderive:('a -> 'b M.t) -> type_name:string -> 'a list -> Ppxlib.expression M.t

Build the functional encoding of a composite type. Combine the various elements necessary for a functional encoding of a composite type ('a, 'b) {!typ}, in terms its components of type 'a list and the name of the composite type type_name.

This requires a function subderive for deriving the generic representation of the subcomponents, which may run in a monadic context M.t.